Better tmux<->vim navigation, <Space> as leader
C-h, C-j, C-k and C-l are now used to navigate between panes in both
tmux and vim windows. Because of this, vim-submode and entire custom
window submode in vim is not needer anymore, so I removed it. Also,
changed vim leader to <Space> and tmux prefix to C-<Space>, because
it is easier to type and remember.

This change solves problem of inconsistent vim and tmux behaviour
and decreases number of custom keybindings in vim.

Signed-off-by: Tomas Slusny <slusnucky@gmail.com>